+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## vNext
+## Unreleased
 
 ### Changes
 
 - Consolidate logic for resolving hub (fixes bug "SENTRY_DSN is not defined") (#795) @Tyrrrz
+- Add SetFingerprint overload that takes `params string[]` (#796) @Tyrrrz
 - Create spans for outgoing HTTP requests (#802) @Tyrrrz
 
 ## 3.0.3